requirements.txt
google-generativeai
pandas
inflect